Move ginkgo_executors.hpp into kernels/splines (#371)
