+++ include/ddc/ddc.hpp
@@ -42,7 +42,3 @@
 #if defined(DDC_BUILD_PDI_WRAPPER)
 #include "ddc/pdi.hpp"
 #endif
-
-#if ginkgo_AVAIL
-#include "misc/ginkgo_executors.hpp"
-#endif
