+++ include/ddc/kernels/splines/ginkgo_executors.hpp
